fix: renamed few modules and github actions added for ktlint
ALS-1978
