+++ app/build.gradle
@@ -130,12 +130,13 @@ android {
 }
 
 ktlint {
-    android = true // Enable Android-specific linting rules
-    ignoreFailures = false // Fail the build if KtLint finds any issues
+    android = true
+    ignoreFailures = false
     disabledRules = ["max-line-length"]
     reporters {
-        reporter("plain") // Output KtLint results in plain text format
-        reporter("html") // Output KtLint results in HTML format
+        reporter("plain")
+        reporter("html")
+        reporter("checkstyle")
     }
 }
