+++ app/src/androidTest/java/com/aws/amazonlocation/ui/main/ExploreFragmentMapZoomInOutTest.kt
@@ -14,7 +14,7 @@ import com.aws.amazonlocation.actions.pinchIn
 import com.aws.amazonlocation.actions.pinchOut
 import com.aws.amazonlocation.checkLocationPermission
 import com.aws.amazonlocation.di.AppModule
-import com.aws.amazonlocation.utils.retry_rule.Retry
+import com.aws.amazonlocation.utils.retryRule.Retry
 import com.aws.amazonlocation.waitForView
 import com.aws.amazonlocation.waitUntil
 import dagger.hilt.android.testing.HiltAndroidTest
