+++ app/src/androidTest/java/com/aws/amazonlocation/utils/screenshotRule/ScreenshotTakingRule.kt
@@ -1,4 +1,4 @@
-package com.aws.amazonlocation.utils.screenshot_rule
+package com.aws.amazonlocation.utils.screenshotRule
 
 import org.junit.rules.TestWatcher
 import org.junit.runner.Description
