--- app/src/androidTest/java/com/aws/amazonlocation/utils/screenshot_rule/MyScreenCaptureProcessor.kt
@@ -1,21 +1,0 @@
-package com.aws.amazonlocation.utils.screenshot_rule
-
-import android.os.Environment.DIRECTORY_PICTURES
-import android.os.Environment.getExternalStoragePublicDirectory
-import androidx.test.runner.screenshot.BasicScreenCaptureProcessor
-import java.io.File
-
-class MyScreenCaptureProcessor(parentFolderPath: String) : BasicScreenCaptureProcessor() {
-
-    init {
-        this.mDefaultScreenshotPath = File(
-            File(
-                getExternalStoragePublicDirectory(DIRECTORY_PICTURES),
-                "my_app_folder"
-            ).absolutePath,
-            "screenshots/$parentFolderPath"
-        )
-    }
-
-    override fun getFilename(prefix: String): String = prefix
-}
