--- app/src/androidTest/java/com/aws/amazonlocation/utils/screenshot_rule/ScreenshotTaker.kt
@@ -1,22 +1,0 @@
-package com.aws.amazonlocation.utils.screenshot_rule
-
-import android.util.Log
-import androidx.test.runner.screenshot.Screenshot
-import java.io.IOException
-
-fun takeScreenshot(parentFolderPath: String = "", screenShotName: String) {
-    Log.d(TAG, "Taking screenshot of '$screenShotName'")
-    val screenCapture = Screenshot.capture()
-    val processors = setOf(MyScreenCaptureProcessor(parentFolderPath))
-    try {
-        screenCapture.apply {
-            name = screenShotName
-            process(processors)
-        }
-        Log.d(TAG, "Screenshot taken")
-    } catch (ex: IOException) {
-        Log.e(TAG, "Could not take a screenshot", ex)
-    }
-}
-
-private const val TAG = "Screenshots"
