+++ app/src/main/java/com/aws/amazonlocation/ui/main/MainActivity.kt
@@ -39,7 +39,7 @@ import com.aws.amazonlocation.domain.`interface`.SignInConnectInterface
 import com.aws.amazonlocation.domain.`interface`.SignInRequiredInterface
 import com.aws.amazonlocation.ui.base.BaseActivity
 import com.aws.amazonlocation.ui.main.explore.ExploreFragment
-import com.aws.amazonlocation.ui.main.map_style.MapStyleFragment
+import com.aws.amazonlocation.ui.main.mapStyle.MapStyleFragment
 import com.aws.amazonlocation.ui.main.setting.AWSCloudInformationFragment
 import com.aws.amazonlocation.ui.main.setting.SettingFragment
 import com.aws.amazonlocation.ui.main.simulation.SimulationUtils
