+++ app/src/main/java/com/aws/amazonlocation/ui/main/attribution/AttributionFragment.kt
@@ -13,7 +13,7 @@ import com.aws.amazonlocation.BuildConfig
 import com.aws.amazonlocation.R
 import com.aws.amazonlocation.databinding.FragmentAttributionBinding
 import com.aws.amazonlocation.ui.base.BaseFragment
-import com.aws.amazonlocation.ui.main.web_view.WebViewActivity
+import com.aws.amazonlocation.ui.main.webView.WebViewActivity
 import com.aws.amazonlocation.utils.KEY_URL
 import com.aws.amazonlocation.utils.MAP_STYLE_ATTRIBUTION
 import com.aws.amazonlocation.utils.attributionPattern
