+++ app/src/main/java/com/aws/amazonlocation/ui/main/more/AboutFragment.kt
@@ -15,8 +15,8 @@ import com.aws.amazonlocation.ui.base.BaseFragment
 import com.aws.amazonlocation.ui.main.MainActivity
 import com.aws.amazonlocation.ui.main.about.VersionFragment
 import com.aws.amazonlocation.ui.main.attribution.AttributionFragment
-import com.aws.amazonlocation.ui.main.terms_condition.TermsAndConditionFragment
-import com.aws.amazonlocation.ui.main.web_view.WebViewActivity
+import com.aws.amazonlocation.ui.main.termsCondition.TermsAndConditionFragment
+import com.aws.amazonlocation.ui.main.webView.WebViewActivity
 import com.aws.amazonlocation.utils.AnalyticsAttributeValue
 import com.aws.amazonlocation.utils.KEY_URL
