+++ app/src/main/java/com/aws/amazonlocation/ui/main/signin/CloudFormationBottomSheetFragment.kt
@@ -20,7 +20,7 @@ import com.aws.amazonlocation.data.enum.TabEnum
 import com.aws.amazonlocation.databinding.BottomSheetCloudFormationBinding
 import com.aws.amazonlocation.domain.`interface`.CloudFormationInterface
 import com.aws.amazonlocation.ui.main.MainActivity
-import com.aws.amazonlocation.ui.main.web_view.WebViewActivity
+import com.aws.amazonlocation.ui.main.webView.WebViewActivity
 import com.aws.amazonlocation.utils.AnalyticsAttribute
 import com.aws.amazonlocation.utils.AnalyticsAttributeValue
 import com.aws.amazonlocation.utils.EventType
