+++ app/src/main/java/com/aws/amazonlocation/utils/geofenceHelper/turf/TurfException.java
@@ -1,4 +1,4 @@
-package com.aws.amazonlocation.utils.geofence_helper.turf;
+package com.aws.amazonlocation.utils.geofenceHelper.turf;
 
 /**
  * This indicates conditions that a reasonable application might want to catch.
