+++ app/src/main/java/com/aws/amazonlocation/utils/geofenceHelper/turf/TurfMeta.java
@@ -1,4 +1,4 @@
-package com.aws.amazonlocation.utils.geofence_helper.turf;
+package com.aws.amazonlocation.utils.geofenceHelper.turf;
 
 
 import com.aws.amazonlocation.utils.ExcludeFromJacocoGeneratedReport;
