+++ app/src/main/res/layout-sw720dp/fragment_map_style.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:id="@+id/cl_map_style_main"
     android:fillViewport="true"
-    tools:context=".ui.main.map_style.MapStyleFragment">
+    tools:context=".ui.main.mapStyle.MapStyleFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
