+++ app/src/main/res/layout/fragment_map_style.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/cl_map_style_main"
-    tools:context=".ui.main.map_style.MapStyleFragment">
+    tools:context=".ui.main.mapStyle.MapStyleFragment">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/iv_map_style_back"
