+++ app/src/test/java/com/aws/amazonlocation/utils/turf/TurfMetaCoordsMultiPolygonTest.kt
@@ -2,7 +2,7 @@ package com.aws.amazonlocation.utils.turf
 
 import com.aws.amazonlocation.mock.TEST_DATA_LAT
 import com.aws.amazonlocation.mock.TEST_DATA_LNG
-import com.aws.amazonlocation.utils.geofence_helper.turf.TurfMeta
+import com.aws.amazonlocation.utils.geofenceHelper.turf.TurfMeta
 import org.junit.Assert
 import org.junit.Test
 import org.junit.runner.RunWith
