+++ app/src/test/java/com/aws/amazonlocation/viewmodel/mapstyle/MapStyleVMSetLanguageListData.kt
@@ -5,7 +5,7 @@ import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.aws.amazonlocation.BaseTest
 import com.aws.amazonlocation.data.response.LanguageData
 import com.aws.amazonlocation.mock.TEST_FAILED_DUE_TO_INCORRECT_NO_OF_LANGUAGE_LOADED
-import com.aws.amazonlocation.ui.main.map_style.MapStyleViewModel
+import com.aws.amazonlocation.ui.main.mapStyle.MapStyleViewModel
 import kotlinx.coroutines.test.runTest
 import org.junit.Assert
 import org.junit.Rule
